requirements.txt
openai>=1.0.0
requests>=2.25.0
python-dotenv>=0.15.0
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0 